+++ mayan/apps/document_comments/apps.py
@@ -7,7 +7,7 @@ from mayan.apps.common.menus import (
     menu_facet, menu_list_facet, menu_object, menu_secondary
 )
 from mayan.apps.documents.search import document_page_search, document_search
-from mayan.apps.events.classes import ModelEventType
+from mayan.apps.events.classes import EventModelRegistry, ModelEventType
 from mayan.apps.events.links import (
     link_events_for_object, link_object_event_types_user_subcriptions_list
 )
@@ -38,13 +38,14 @@ class DocumentCommentsApp(MayanAppConfig):
 
     def ready(self):
         super(DocumentCommentsApp, self).ready()
-        from actstream import registry
 
         Document = apps.get_model(
             app_label='documents', model_name='Document'
         )
 
         Comment = self.get_model(model_name='Comment')
+
+        EventModelRegistry.register(model=Comment)
 
         ModelEventType.register(
             model=Comment, event_types=(
@@ -113,5 +114,3 @@ class DocumentCommentsApp(MayanAppConfig):
         menu_object.bind_links(
             links=(link_comment_delete, link_comment_edit), sources=(Comment,)
         )
-
-        registry.register(Comment)